Improve Quarkus configuration tests execution
