+++ quarkus/runtime/src/test/java/org/keycloak/quarkus/runtime/configuration/test/ManagementConfigurationTest.java
@@ -24,7 +24,7 @@ import java.util.Map;
 import static org.hamcrest.CoreMatchers.is;
 import static org.hamcrest.MatcherAssert.assertThat;
 
-public class ManagementConfigurationTest extends ConfigurationTest {
+public class ManagementConfigurationTest extends AbstractConfigurationTest {
 
     @Test
     public void managementDefaults() {
